refactor(post, page): move the paging core code to the `service`
