+++ src/modules/post/post.service.ts
@@ -6,7 +6,7 @@ import {
 } from "@nestjs/common";
 import { PostModel } from "./post.model";
 import { InjectModel } from "~/transformers/model.transformer";
-import { AggregatePaginateModel, Document } from "mongoose";
+import { AggregatePaginateModel, Document, PipelineStage } from "mongoose";
 import { CategoryService } from "../category/category.service";
 import slugify from "slugify";
 import { isDefined } from "class-validator";
@@ -15,6 +15,8 @@ import { BusinessException } from "~/common/exceptions/business.excpetion";
 import { ErrorCodeEnum } from "~/constants/error-code.constant";
 import { CommentModel, CommentType } from "../comments/comments.model";
 import { ImageService } from "~/processors/helper/helper.image.service";
+import { PagerDto } from "~/shared/dto/pager.dto";
+import { addYearCondition } from "~/transformers/db-query.transformer";
 
 @Injectable()
 export class PostService {
@@ -30,6 +32,122 @@ export class PostService {
   ) {}
   get model() {
     return this.postModel;
+  }
+
+  async aggregatePaginate(query: PagerDto, isMaster = false) {
+    const { size, select, page, year, sortBy, sortOrder } = query;
+    return this.model.aggregatePaginate(
+      this.model.aggregate(
+        [
+          {
+            $match: {
+              ...addYearCondition(year),
+            },
+          },
+          // @see https://stackoverflow.com/questions/54810712/mongodb-sort-by-field-a-if-field-b-null-otherwise-sort-by-field-c
+          {
+            $addFields: {
+              sortField: {
+                // create a new field called "sortField"
+                $cond: {
+                  // and assign a value that depends on
+                  if: { $ne: ["$pin", null] }, // whether "b" is not null
+                  then: "$pinOrder", // in which case our field shall hold the value of "a"
+                  else: "$$REMOVE",
+                },
+              },
+            },
+          },
+          // if not master, only show published posts
+          !isMaster && {
+            $match: {
+              // match the condition
+              hide: { $ne: true }, // $ne: not equal
+            },
+          },
+          // 如果不是master，并且password不为空，则将text,summary修改
+          !isMaster && {
+            $set: {
+              // set the field to a new value
+              summary: {
+                $cond: {
+                  if: { $ne: ["$password", null] }, // if "password" is not null
+                  then: { $concat: ["内容已被隐藏，请输入密码"] }, // then the value of "内容已被隐藏"
+                  else: "$title", // otherwise, use the original title
+                }, // $concat: 用于拼接字符串
+              },
+              text: {
+                $cond: {
+                  // 如果密码字段不为空，且isMaster为false，则不显示
+                  if: {
+                    $ne: ["$password", null],
+                  }, // whether "b" is not null
+                  then: { $concat: ["内容已被隐藏，请输入密码"] },
+                  else: "$text",
+                },
+              },
+            },
+          },
+          !isMaster && {
+            // if not master, only show usual fields
+            $project: {
+              hide: 0,
+              password: 0,
+              rss: 0,
+            },
+          },
+          {
+            $sort: sortBy
+              ? {
+                  [sortBy]: sortOrder as any,
+                }
+              : {
+                  sortField: -1, // sort by our computed field
+                  pin: -1,
+                  created: -1, // and then by the "created" field
+                },
+          },
+          {
+            $project: {
+              // project the fields we want to keep
+              sortField: 0, // remove "sort" field if needed
+            },
+          },
+          select && {
+            $project: {
+              ...(select?.split(" ").reduce(
+                (acc, cur) => {
+                  const field = cur.trim();
+                  acc[field] = 1;
+                  return acc;
+                },
+                Object.keys(new PostModel()).map((k) => ({ [k]: 0 }))
+              ) as any),
+            },
+          },
+          {
+            $lookup: {
+              // lookup can be used to join two collections
+              from: "categories", // from the "categories" collection
+              localField: "categoryId", // using the "categoryId" field
+              foreignField: "_id", // from the "categories" collection
+              as: "category", // as the "category" field
+            },
+          },
+          {
+            $unwind: {
+              // unwind 将数组的每个元素解析为单个文档
+              path: "$category", // the path to the array
+              preserveNullAndEmptyArrays: true, // if set to true, MongoDB will still create a document if the array is empty
+            },
+          },
+        ].filter(Boolean) as PipelineStage[]
+      ),
+      {
+        limit: size,
+        page,
+      }
+    );
   }
 
   /**
